Prohibit reacting to onself posts
fix #10

+++ db.go
@@ -107,6 +107,18 @@ func (bot *tBot) hasPostId(messageId int64) bool {
 }
 
 func (bot *tBot) like(postId int64, reactionType int, userId int64, name string) {
+	var authorId int64
+	err := bot.db.QueryRow(`
+		SELECT author_id
+		FROM authors
+		WHERE post_id = ?`, postId).Scan(&authorId)
+	if err != nil {
+		log.Panic(err)
+	}
+	if authorId == userId {
+		log.Printf("Self reaction of <%v> to %v: %v\n", name, postId, reactionType)
+		return
+	}
 	res, err := bot.db.Exec(`
         DELETE FROM likes
         WHERE post_id = ? AND reaction_type = ? AND user_id = ?`,
